+++ src/Chunkyard.Tests/MemoryRepository.cs
@@ -9,18 +9,15 @@ namespace Chunkyard.Tests
         private readonly Dictionary<Uri, byte[]> _valuesByUri;
         private readonly Dictionary<int, byte[]> _valuesByLog;
 
-        public MemoryRepository(Guid? repositoryId = null)
+        public MemoryRepository()
         {
             _valuesByUri = new Dictionary<Uri, byte[]>();
             _valuesByLog = new Dictionary<int, byte[]>();
 
             RepositoryUri = new Uri("in://memory");
-            RepositoryId = repositoryId ?? Guid.NewGuid();
         }
 
         public Uri RepositoryUri { get; }
-
-        public Guid RepositoryId { get; }
 
         public bool StoreValue(Uri contentUri, byte[] value)
         {